const String& 2

+++ src/notification_utils.cpp
@@ -1,22 +1,22 @@
 #include "notification_utils.h"
 #include "configuration.h"
 
-int channel       = 0;
-int resolution    = 8; 
-int pauseDuration = 20;
+uint8_t channel                 = 0;
+uint8_t resolution              = 8; 
+uint8_t pauseDuration           = 20;
 
-int startUpSound[]          = {440, 880, 440, 1760};
-int startUpSoundDuration[]  = {100, 100, 100, 200};
+int     startUpSound[]          = {440, 880, 440, 1760};
+uint8_t startUpSoundDuration[]  = {100, 100, 100, 200};
 
-int shutDownSound[]         = {1720, 880, 400};
-int shutDownSoundDuration[] = {60, 60, 200};
+int     shutDownSound[]         = {1720, 880, 400};
+uint8_t shutDownSoundDuration[] = {60, 60, 200};
 
 extern Configuration    Config;
 extern bool             digirepeaterActive;
 
 namespace NOTIFICATION_Utils {
 
-    void playTone(int frequency, int duration) {
+    void playTone(int frequency, uint8_t duration) {
         ledcSetup(channel, frequency, resolution);
         ledcAttachPin(Config.notification.buzzerPinTone, 0);
         ledcWrite(channel, 128);
